Feature: isolate ontology submission process steps (#132)
* add an abstraction for submission process steps

* extract submission generate_rdf step to a file

* extract submission generate missing labels steps into a file

* extract the submission archiving step into a file

* add  abstraction to diff tool & extract the submission step to a file

* extract the submission metrics generation step to a file

* extract the submission properties indexation step into a file

* extract the submission terms indexation step into a file

* move the extract metadata concern to submission process step file

* extract the submission generate obsolete classes step from generate rdf

* add the global submission process that call the sub-steps
