+++ lib/ontologies_linked_data/models/ontology_submission.rb
@@ -12,8 +12,8 @@ module LinkedData
 
     class OntologySubmission < LinkedData::Models::Base
 
-      include LinkedData::Concerns::OntologySubmission::MetadataExtractor
       include LinkedData::Concerns::OntologySubmission::IndexAllData
+      include LinkedData::Concerns::SubmissionProcessable
       include LinkedData::Concerns::OntologySubmission::Validators
       include LinkedData::Concerns::OntologySubmission::UpdateCallbacks
       extend LinkedData::Concerns::OntologySubmission::DefaultCallbacks
@@ -21,10 +21,8 @@ module LinkedData
       include SKOS::ConceptSchemes
       include SKOS::RootsFetcher
 
-      FILES_TO_DELETE = ['labels.ttl', 'mappings.ttl', 'obsolete.ttl', 'owlapi.xrdf', 'errors.log']
-      FOLDERS_TO_DELETE = ['unzipped']
+
       FLAT_ROOTS_LIMIT = 1000
-      FILE_SIZE_ZIPPING_THRESHOLD = 100 * 1024 * 1024 # 100MB
 
       model :ontology_submission, scheme: File.join(__dir__, '../../../config/schemes/ontology_submission.yml'),
             name_with: ->(s) { submission_id_generator(s) }
@@ -472,54 +470,8 @@ module LinkedData
         zip_dst
       end
 
-      def delete_old_submission_files
-        path_to_repo = data_folder
-        submission_files = FILES_TO_DELETE.map { |f| File.join(path_to_repo, f) }
-        submission_files.push(csv_path)
-        submission_files.push(parsing_log_path) unless parsing_log_path.nil?
-        FileUtils.rm(submission_files, force: true)
 
-        submission_folders = FOLDERS_TO_DELETE.map { |f| File.join(path_to_repo, f) }
-        submission_folders.each { |d| FileUtils.remove_dir(d) if File.directory?(d) }
-      end
 
-      def zip_submission_uploaded_file
-        self.bring(:uploadFilePath) if self.bring?(:uploadFilePath)
-
-        return self.uploadFilePath if zipped?
-        return self.uploadFilePath if self.uploadFilePath.nil? || self.uploadFilePath.empty?
-
-        return self.uploadFilePath if File.size(self.uploadFilePath) < FILE_SIZE_ZIPPING_THRESHOLD
-
-        old_path = self.uploadFilePath
-        new_path = Utils::FileHelpers.zip_file(old_path)
-        FileUtils.rm(old_path, force: true)
-        new_path
-      end
-
-      # accepts another submission in 'older' (it should be an 'older' ontology version)
-      def diff(logger, older)
-        begin
-          bring_remaining
-          bring :diffFilePath if bring? :diffFilePath
-          older.bring :uploadFilePath if older.bring? :uploadFilePath
-
-          LinkedData::Diff.logger = logger
-          bubastis = LinkedData::Diff::BubastisDiffCommand.new(
-            File.expand_path(older.master_file_path),
-            File.expand_path(self.master_file_path),
-            data_folder
-          )
-          self.diffFilePath = bubastis.diff
-          save
-          logger.info("Bubastis diff generated successfully for #{self.id}")
-          logger.flush
-        rescue Exception => e
-          logger.error("Bubastis diff for #{self.id} failed - #{e.class}: #{e.message}")
-          logger.flush
-          raise e
-        end
-      end
 
       def class_count(logger = nil)
         logger ||= LinkedData::Parser.logger || Logger.new($stderr)
@@ -578,372 +530,6 @@ module LinkedData
         metrics
       end
 
-      def generate_metrics_file(class_count, indiv_count, prop_count)
-        CSV.open(self.metrics_path, "wb") do |csv|
-          csv << ["Class Count", "Individual Count", "Property Count"]
-          csv << [class_count, indiv_count, prop_count]
-        end
-      end
-
-      def generate_metrics_file2(class_count, indiv_count, prop_count, max_depth)
-        CSV.open(self.metrics_path, "wb") do |csv|
-          csv << ["Class Count", "Individual Count", "Property Count", "Max Depth"]
-          csv << [class_count, indiv_count, prop_count, max_depth]
-        end
-      end
-
-      def generate_umls_metrics_file(tr_file_path = nil)
-        tr_file_path ||= self.triples_file_path
-        class_count = 0
-        indiv_count = 0
-        prop_count = 0
-
-        File.foreach(tr_file_path) do |line|
-          class_count += 1 if line =~ /owl:Class/
-          indiv_count += 1 if line =~ /owl:NamedIndividual/
-          prop_count += 1 if line =~ /owl:ObjectProperty/
-          prop_count += 1 if line =~ /owl:DatatypeProperty/
-        end
-        self.generate_metrics_file(class_count, indiv_count, prop_count)
-      end
-
-      def generate_rdf(logger, reasoning: true)
-        mime_type = nil
-
-        if self.hasOntologyLanguage.umls?
-          triples_file_path = self.triples_file_path
-          logger.info("Using UMLS turtle file found, skipping OWLAPI parse")
-          logger.flush
-          mime_type = LinkedData::MediaTypes.media_type_from_base(LinkedData::MediaTypes::TURTLE)
-          generate_umls_metrics_file(triples_file_path)
-        else
-          output_rdf = self.rdf_path
-
-          if File.exist?(output_rdf)
-            logger.info("deleting old owlapi.xrdf ..")
-            deleted = FileUtils.rm(output_rdf)
-
-            if deleted.length > 0
-              logger.info("deleted")
-            else
-              logger.info("error deleting owlapi.rdf")
-            end
-          end
-          owlapi = owlapi_parser(logger: logger)
-
-          owlapi.disable_reasoner if !reasoning
-          triples_file_path, missing_imports = owlapi.parse
-
-          if missing_imports && missing_imports.length > 0
-            self.missingImports = missing_imports
-
-            missing_imports.each do |imp|
-              logger.info("OWL_IMPORT_MISSING: #{imp}")
-            end
-          else
-            self.missingImports = nil
-          end
-          logger.flush
-        end
-        delete_and_append(triples_file_path, logger, mime_type)
-      end
-
-      def process_callbacks(logger, callbacks, action_name, &block)
-        callbacks.delete_if do |_, callback|
-          begin
-            if callback[action_name]
-              callable = self.method(callback[action_name])
-              yield(callable, callback)
-            end
-            false
-          rescue Exception => e
-            logger.error("#{e.class}: #{e.message}\n#{e.backtrace.join("\n\t")}")
-            logger.flush
-
-            if callback[:status]
-              add_submission_status(callback[:status].get_error_status)
-              self.save
-            end
-
-            # halt the entire processing if :required is set to true
-            raise e if callback[:required]
-            # continue processing of other callbacks, but not this one
-            true
-          end
-        end
-      end
-
-      def loop_classes(logger, raw_paging, callbacks)
-        page = 1
-        size = 2500
-        count_classes = 0
-        acr = self.id.to_s.split("/")[-1]
-        operations = callbacks.values.map { |v| v[:op_name] }.join(", ")
-
-        time = Benchmark.realtime do
-          paging = raw_paging.page(page, size)
-          cls_count_set = false
-          cls_count = class_count(logger)
-
-          if cls_count > -1
-            # prevent a COUNT SPARQL query if possible
-            paging.page_count_set(cls_count)
-            cls_count_set = true
-          else
-            cls_count = 0
-          end
-
-          iterate_classes = false
-          # 1. init artifacts hash if not explicitly passed in the callback
-          # 2. determine if class-level iteration is required
-          callbacks.each { |_, callback| callback[:artifacts] ||= {};
-          if callback[:caller_on_each]
-            iterate_classes = true
-          end }
-
-          process_callbacks(logger, callbacks, :caller_on_pre) {
-            |callable, callback| callable.call(callback[:artifacts], logger, paging) }
-
-          page_len = -1
-          prev_page_len = -1
-
-          begin
-            t0 = Time.now
-            page_classes = paging.page(page, size).all
-            total_pages = page_classes.total_pages
-            page_len = page_classes.length
-
-            # nothing retrieved even though we're expecting more records
-            if total_pages > 0 && page_classes.empty? && (prev_page_len == -1 || prev_page_len == size)
-              j = 0
-              num_calls = LinkedData.settings.num_retries_4store
-
-              while page_classes.empty? && j < num_calls do
-                j += 1
-                logger.error("Empty page encountered. Retrying #{j} times...")
-                sleep(2)
-                page_classes = paging.page(page, size).all
-                unless page_classes.empty?
-                  logger.info("Success retrieving a page of #{page_classes.length} classes after retrying #{j} times...")
-                end
-              end
-
-              if page_classes.empty?
-                msg = "Empty page #{page} of #{total_pages} persisted after retrying #{j} times. #{operations} of #{acr} aborted..."
-                logger.error(msg)
-                raise msg
-              end
-            end
-
-            if page_classes.empty?
-              if total_pages > 0
-                logger.info("The number of pages reported for #{acr} - #{total_pages} is higher than expected #{page - 1}. Completing #{operations}...")
-              else
-                logger.info("Ontology #{acr} contains #{total_pages} pages...")
-              end
-              break
-            end
-
-            prev_page_len = page_len
-            logger.info("#{acr}: page #{page} of #{total_pages} - #{page_len} ontology terms retrieved in #{Time.now - t0} sec.")
-            logger.flush
-            count_classes += page_classes.length
-
-            process_callbacks(logger, callbacks, :caller_on_pre_page) {
-              |callable, callback| callable.call(callback[:artifacts], logger, paging, page_classes, page) }
-
-            page_classes.each { |c|
-              # For real this is calling "generate_missing_labels_each". Is it that hard to be clear in your code?
-              # It is unreadable, not stable and not powerful. What did you want to do?
-              process_callbacks(logger, callbacks, :caller_on_each) {
-                |callable, callback| callable.call(callback[:artifacts], logger, paging, page_classes, page, c) }
-            } if iterate_classes
-
-            process_callbacks(logger, callbacks, :caller_on_post_page) {
-              |callable, callback| callable.call(callback[:artifacts], logger, paging, page_classes, page) }
-            cls_count += page_classes.length unless cls_count_set
-
-            page = page_classes.next? ? page + 1 : nil
-          end while !page.nil?
-
-          callbacks.each { |_, callback| callback[:artifacts][:count_classes] = cls_count }
-          process_callbacks(logger, callbacks, :caller_on_post) {
-            |callable, callback| callable.call(callback[:artifacts], logger, paging) }
-        end
-
-        logger.info("Completed #{operations}: #{acr} in #{time} sec. #{count_classes} classes.")
-        logger.flush
-
-        # set the status on actions that have completed successfully
-        callbacks.each do |_, callback|
-          if callback[:status]
-            add_submission_status(callback[:status])
-            self.save
-          end
-        end
-      end
-
-      def generate_missing_labels_pre(artifacts = {}, logger, paging)
-        file_path = artifacts[:file_path]
-        artifacts[:save_in_file] = File.join(File.dirname(file_path), "labels.ttl")
-        artifacts[:save_in_file_mappings] = File.join(File.dirname(file_path), "mappings.ttl")
-        property_triples = LinkedData::Utils::Triples.rdf_for_custom_properties(self)
-        Goo.sparql_data_client.append_triples(self.id, property_triples, mime_type = "application/x-turtle")
-        fsave = File.open(artifacts[:save_in_file], "w")
-        fsave.write(property_triples)
-        fsave_mappings = File.open(artifacts[:save_in_file_mappings], "w")
-        artifacts[:fsave] = fsave
-        artifacts[:fsave_mappings] = fsave_mappings
-      end
-
-      def generate_missing_labels_pre_page(artifacts = {}, logger, paging, page_classes, page)
-        artifacts[:label_triples] = []
-        artifacts[:mapping_triples] = []
-      end
-
-      # Generate labels when no label found in the prefLabel attribute (it checks rdfs:label and take label from the URI if nothing else found)
-      def generate_missing_labels_each(artifacts = {}, logger, paging, page_classes, page, c)
-        prefLabel = nil
-
-        if c.prefLabel.nil?
-          begin
-            # in case there is no skos:prefLabel or rdfs:label from our main_lang
-            rdfs_labels = c.label
-
-            if rdfs_labels && rdfs_labels.length > 1 && c.synonym.length > 0
-              rdfs_labels = (Set.new(c.label) - Set.new(c.synonym)).to_a.first
-
-              rdfs_labels = c.label if rdfs_labels.nil? || rdfs_labels.length == 0
-            end
-
-            rdfs_labels = [rdfs_labels] if rdfs_labels and not (rdfs_labels.instance_of? Array)
-            label = nil
-
-            if rdfs_labels && rdfs_labels.length > 0
-              label = rdfs_labels[0]
-            else
-              # If no label found, we take the last fragment of the URI
-              label = LinkedData::Utils::Triples.last_iri_fragment c.id.to_s
-            end
-          rescue Goo::Base::AttributeNotLoaded => e
-            label = LinkedData::Utils::Triples.last_iri_fragment c.id.to_s
-          end
-          artifacts[:label_triples] << LinkedData::Utils::Triples.label_for_class_triple(
-            c.id, Goo.vocabulary(:metadata_def)[:prefLabel], label)
-          prefLabel = label
-        else
-          prefLabel = c.prefLabel
-        end
-
-        if self.ontology.viewOf.nil?
-          loomLabel = OntologySubmission.loom_transform_literal(prefLabel.to_s)
-
-          if loomLabel.length > 2
-            artifacts[:mapping_triples] << LinkedData::Utils::Triples.loom_mapping_triple(
-              c.id, Goo.vocabulary(:metadata_def)[:mappingLoom], loomLabel)
-          end
-          artifacts[:mapping_triples] << LinkedData::Utils::Triples.uri_mapping_triple(
-            c.id, Goo.vocabulary(:metadata_def)[:mappingSameURI], c.id)
-        end
-      end
-
-      def generate_missing_labels_post_page(artifacts = {}, logger, paging, page_classes, page)
-        rest_mappings = LinkedData::Mappings.migrate_rest_mappings(self.ontology.acronym)
-        artifacts[:mapping_triples].concat(rest_mappings)
-
-        if artifacts[:label_triples].length > 0
-          logger.info("Asserting #{artifacts[:label_triples].length} labels in " +
-                        "#{self.id.to_ntriples}")
-          logger.flush
-          artifacts[:label_triples] = artifacts[:label_triples].join("\n")
-          artifacts[:fsave].write(artifacts[:label_triples])
-          t0 = Time.now
-          Goo.sparql_data_client.append_triples(self.id, artifacts[:label_triples], mime_type = "application/x-turtle")
-          t1 = Time.now
-          logger.info("Labels asserted in #{t1 - t0} sec.")
-          logger.flush
-        else
-          logger.info("No labels generated in page #{page}.")
-          logger.flush
-        end
-
-        if artifacts[:mapping_triples].length > 0
-          logger.info("Asserting #{artifacts[:mapping_triples].length} mappings in " +
-                        "#{self.id.to_ntriples}")
-          logger.flush
-          artifacts[:mapping_triples] = artifacts[:mapping_triples].join("\n")
-          artifacts[:fsave_mappings].write(artifacts[:mapping_triples])
-
-          t0 = Time.now
-          Goo.sparql_data_client.append_triples(self.id, artifacts[:mapping_triples], mime_type = "application/x-turtle")
-          t1 = Time.now
-          logger.info("Mapping labels asserted in #{t1 - t0} sec.")
-          logger.flush
-        end
-      end
-
-      def generate_missing_labels_post(artifacts = {}, logger, paging)
-        logger.info("end generate_missing_labels traversed #{artifacts[:count_classes]} classes")
-        logger.info("Saved generated labels in #{artifacts[:save_in_file]}")
-        artifacts[:fsave].close()
-        artifacts[:fsave_mappings].close()
-        logger.flush
-      end
-
-      def generate_obsolete_classes(logger, file_path)
-        self.bring(:obsoleteProperty) if self.bring?(:obsoleteProperty)
-        self.bring(:obsoleteParent) if self.bring?(:obsoleteParent)
-        classes_deprecated = []
-        if self.obsoleteProperty &&
-          self.obsoleteProperty.to_s != "http://www.w3.org/2002/07/owl#deprecated"
-
-          predicate_obsolete = RDF::URI.new(self.obsoleteProperty.to_s)
-          query_obsolete_predicate = <<eos
-SELECT ?class_id ?deprecated
-FROM #{self.id.to_ntriples}
-WHERE { ?class_id #{predicate_obsolete.to_ntriples} ?deprecated . }
-eos
-          Goo.sparql_query_client.query(query_obsolete_predicate).each_solution do |sol|
-            classes_deprecated << sol[:class_id].to_s unless ["0", "false"].include? sol[:deprecated].to_s
-          end
-          logger.info("Obsolete found #{classes_deprecated.length} for property #{self.obsoleteProperty.to_s}")
-        end
-        if self.obsoleteParent.nil?
-          # try to find oboInOWL obsolete.
-          obo_in_owl_obsolete_class = LinkedData::Models::Class
-                                        .find(LinkedData::Utils::Triples.obo_in_owl_obsolete_uri)
-                                        .in(self).first
-          self.obsoleteParent = LinkedData::Utils::Triples.obo_in_owl_obsolete_uri if obo_in_owl_obsolete_class
-        end
-        if self.obsoleteParent
-          class_obsolete_parent = LinkedData::Models::Class
-                                    .find(self.obsoleteParent)
-                                    .in(self).first
-          if class_obsolete_parent
-            descendents_obsolete = class_obsolete_parent.descendants
-            logger.info("Found #{descendents_obsolete.length} descendents of obsolete root #{self.obsoleteParent.to_s}")
-            descendents_obsolete.each do |obs|
-              classes_deprecated << obs.id
-            end
-          else
-            logger.error("Submission #{self.id.to_s} obsoleteParent #{self.obsoleteParent.to_s} not found")
-          end
-        end
-        if classes_deprecated.length > 0
-          classes_deprecated.uniq!
-          logger.info("Asserting owl:deprecated statement for #{classes_deprecated} classes")
-          save_in_file = File.join(File.dirname(file_path), "obsolete.ttl")
-          fsave = File.open(save_in_file, "w")
-          classes_deprecated.each do |class_id|
-            fsave.write(LinkedData::Utils::Triples.obselete_class_triple(class_id) + "\n")
-          end
-          fsave.close()
-          result = Goo.sparql_data_client.append_triples_from_file(
-            self.id,
-            save_in_file,
-            mime_type = "application/x-turtle")
-        end
-      end
 
       def add_submission_status(status)
         valid = status.is_a?(LinkedData::Models::SubmissionStatus)
@@ -1027,504 +613,6 @@ eos
         return ready?(status: [:archived])
       end
 
-      def archive_submission
-        self.submissionStatus = nil
-        status = LinkedData::Models::SubmissionStatus.find("ARCHIVED").first
-        add_submission_status(status)
-        unindex
-        # Delete everything except for original ontology file.
-        ontology.bring(:submissions)
-        submissions = ontology.submissions
-        unless submissions.nil?
-          submissions.each { |s| s.bring(:submissionId) }
-          submission = submissions.sort { |a, b| b.submissionId <=> a.submissionId }[0]
-          # Don't perform deletion if this is the most recent submission.
-          if self.submissionId < submission.submissionId
-            delete_old_submission_files
-            self.uploadFilePath = zip_submission_uploaded_file
-          end
-        end
-      end
-
-      ################################################################
-      # Possible options with their defaults:
-      #   process_rdf       = false
-      #   index_search      = false
-      #   index_properties  = false
-      #   index_commit      = false
-      #   run_metrics       = false
-      #   reasoning         = false
-      #   diff              = false
-      #   archive           = false
-      #   if no options passed, ALL actions, except for archive = true
-      ################################################################
-      def process_submission(logger, options = {})
-        # Wrap the whole process so we can email results
-        begin
-          process_rdf = false
-          extract_metadata = false
-          index_search = false
-          index_properties = false
-          index_commit = false
-          run_metrics = false
-          reasoning = false
-          diff = false
-          archive = false
-
-          if options.empty?
-            process_rdf = true
-            generate_missing_labels = true
-            extract_metadata = true
-            index_search = true
-            index_properties = true
-            index_commit = true
-            run_metrics = true
-            reasoning = true
-            diff = true
-            archive = false
-          else
-            process_rdf = options[:process_rdf] == true ? true : false
-            generate_missing_labels = options[:generate_missing_labels].nil? ? process_rdf : options[:generate_missing_labels]
-            extract_metadata = options[:extract_metadata].nil? ? process_rdf : options[:extract_metadata]
-            index_search = options[:index_search] == true ? true : false
-            index_properties = options[:index_properties] == true ? true : false
-            index_commit = options[:index_commit] == true ? true : false
-            run_metrics = options[:run_metrics] == true ? true : false
-
-            if !process_rdf || options[:reasoning] == false
-              reasoning = false
-            else
-              reasoning = true
-            end
-
-            if (!index_search && !index_properties) || options[:index_commit] == false
-              index_commit = false
-            else
-              index_commit = true
-            end
-
-            diff = options[:diff] == true ? true : false
-            archive = options[:archive] == true ? true : false
-          end
-
-          self.bring_remaining
-          self.ontology.bring_remaining
-
-          logger.info("Starting to process #{self.ontology.acronym}/submissions/#{self.submissionId}")
-          logger.flush
-          LinkedData::Parser.logger = logger
-          status = nil
-
-          if archive
-            archive_submission
-          else
-            if process_rdf
-              # Remove processing status types before starting RDF parsing etc.
-              self.submissionStatus = nil
-              status = LinkedData::Models::SubmissionStatus.find("UPLOADED").first
-              add_submission_status(status)
-              self.save
-
-              # Parse RDF
-              begin
-                if not self.valid?
-                  error = "Submission is not valid, it cannot be processed. Check errors."
-                  raise ArgumentError, error
-                end
-                if not self.uploadFilePath
-                  error = "Submission is missing an ontology file, cannot parse."
-                  raise ArgumentError, error
-                end
-                status = LinkedData::Models::SubmissionStatus.find("RDF").first
-                remove_submission_status(status) # remove RDF status before starting
-                generate_rdf(logger, reasoning: reasoning)
-
-                add_submission_status(status)
-                self.save
-              rescue Exception => e
-                logger.error("#{self.errors}")
-                logger.error("#{e.class}: #{e.message}\n#{e.backtrace.join("\n\t")}")
-                logger.flush
-                add_submission_status(status.get_error_status)
-                self.save
-                # If RDF generation fails, no point of continuing
-                raise e
-              end
-            end
-
-            extract_metadata(logger, options[:params], heavy_extraction: extract_metadata)
-
-            if process_rdf && generate_missing_labels
-              file_path = self.uploadFilePath
-              callbacks = {
-                missing_labels: {
-                  op_name: "Missing Labels Generation",
-                  required: true,
-                  status: LinkedData::Models::SubmissionStatus.find("RDF_LABELS").first,
-                  artifacts: {
-                    file_path: file_path
-                  },
-                  caller_on_pre: :generate_missing_labels_pre,
-                  caller_on_pre_page: :generate_missing_labels_pre_page,
-                  caller_on_each: :generate_missing_labels_each,
-                  caller_on_post_page: :generate_missing_labels_post_page,
-                  caller_on_post: :generate_missing_labels_post
-                }
-              }
-
-              raw_paging = LinkedData::Models::Class.in(self).include(:prefLabel, :synonym, :label)
-              loop_classes(logger, raw_paging, callbacks)
-
-              status = LinkedData::Models::SubmissionStatus.find("OBSOLETE").first
-              begin
-                generate_obsolete_classes(logger, file_path)
-                add_submission_status(status)
-                self.save
-              rescue Exception => e
-                logger.error("#{e.class}: #{e.message}\n#{e.backtrace.join("\n\t")}")
-                logger.flush
-                add_submission_status(status.get_error_status)
-                self.save
-                # if obsolete fails the parsing fails
-                raise e
-              end
-            end
-
-            parsed = ready?(status: %i[rdf rdf_labels])
-
-            if index_search
-              unless parsed
-                raise Exception, "The submission #{self.ontology.acronym}/submissions/#{self.submissionId} cannot be indexed because it has not been successfully parsed"
-              end
-              status = LinkedData::Models::SubmissionStatus.find("INDEXED").first
-              begin
-                index_terms(logger, index_commit, false)
-                add_submission_status(status)
-              rescue Exception => e
-                logger.error("#{e.class}: #{e.message}\n#{e.backtrace.join("\n\t")}")
-                logger.flush
-                add_submission_status(status.get_error_status)
-                FileUtils.rm(self.csv_path) if File.file?(self.csv_path)
-              ensure
-                self.save
-              end
-            end
-
-            if index_properties
-              unless parsed
-                raise Exception, "The properties for the submission #{self.ontology.acronym}/submissions/#{self.submissionId} cannot be indexed because it has not been successfully parsed"
-              end
-              status = LinkedData::Models::SubmissionStatus.find("INDEXED_PROPERTIES").first
-              begin
-                index_properties(logger, index_commit, false)
-                add_submission_status(status)
-              rescue Exception => e
-                logger.error("#{e.class}: #{e.message}\n#{e.backtrace.join("\n\t")}")
-                logger.flush
-                add_submission_status(status.get_error_status)
-              ensure
-                self.save
-              end
-            end
-
-            if run_metrics
-              unless parsed
-                raise Exception, "Metrics cannot be generated on the submission #{self.ontology.acronym}/submissions/#{self.submissionId} because it has not been successfully parsed"
-              end
-              status = LinkedData::Models::SubmissionStatus.find("METRICS").first
-              begin
-                process_metrics(logger)
-                add_submission_status(status)
-              rescue Exception => e
-                logger.error("#{e.class}: #{e.message}\n#{e.backtrace.join("\n\t")}")
-                logger.flush
-                self.metrics = nil
-                add_submission_status(status.get_error_status)
-              ensure
-                self.save
-              end
-            end
-
-            if diff
-              status = LinkedData::Models::SubmissionStatus.find("DIFF").first
-              # Get previous submission from ontology.submissions
-              self.ontology.bring(:submissions)
-              submissions = self.ontology.submissions
-
-              unless submissions.nil?
-                submissions.each { |s| s.bring(:submissionId, :diffFilePath) }
-                # Sort submissions in descending order of submissionId, extract last two submissions
-                recent_submissions = submissions.sort { |a, b| b.submissionId <=> a.submissionId }[0..1]
-
-                if recent_submissions.length > 1
-                  # validate that the most recent submission is the current submission
-                  if self.submissionId == recent_submissions.first.submissionId
-                    prev = recent_submissions.last
-
-                    # Ensure that prev is older than the current submission
-                    if self.submissionId > prev.submissionId
-                      # generate a diff
-                      begin
-                        self.diff(logger, prev)
-                        add_submission_status(status)
-                      rescue Exception => e
-                        logger.error("#{e.class}: #{e.message}\n#{e.backtrace.join("\n\t")}")
-                        logger.flush
-                        add_submission_status(status.get_error_status)
-                      ensure
-                        self.save
-                      end
-                    end
-                  end
-                else
-                  logger.info("Bubastis diff: no older submissions available for #{self.id}.")
-                end
-              else
-                logger.info("Bubastis diff: no submissions available for #{self.id}.")
-              end
-            end
-          end
-
-          self.save
-          logger.info("Submission processing of #{self.id} completed successfully")
-          logger.flush
-        ensure
-          # make sure results get emailed
-          begin
-            LinkedData::Utils::Notifications.submission_processed(self)
-          rescue Exception => e
-            logger.error("Email sending failed: #{e.message}\n#{e.backtrace.join("\n\t")}"); logger.flush
-          end
-        end
-        self
-      end
-
-      def process_metrics(logger)
-        metrics = LinkedData::Metrics.metrics_for_submission(self, logger)
-        metrics.id = RDF::URI.new(self.id.to_s + "/metrics")
-        exist_metrics = LinkedData::Models::Metric.find(metrics.id).first
-        exist_metrics.delete if exist_metrics
-        metrics.save
-
-        self.metrics = metrics
-        self
-      end
-
-      def index_terms(logger, commit = true, optimize = true)
-        page = 0
-        size = 1000
-        count_classes = 0
-
-        time = Benchmark.realtime do
-          self.bring(:ontology) if self.bring?(:ontology)
-          self.ontology.bring(:acronym) if self.ontology.bring?(:acronym)
-          self.ontology.bring(:provisionalClasses) if self.ontology.bring?(:provisionalClasses)
-          csv_writer = LinkedData::Utils::OntologyCSVWriter.new
-          csv_writer.open(self.ontology, self.csv_path)
-
-          begin
-            logger.info("Indexing ontology terms: #{self.ontology.acronym}...")
-            t0 = Time.now
-            self.ontology.unindex_all_data(false)
-            logger.info("Removed ontology terms index (#{Time.now - t0}s)"); logger.flush
-
-            paging = LinkedData::Models::Class.in(self).include(:unmapped).aggregate(:count, :children).page(page, size)
-            # a fix for SKOS ontologies, see https://github.com/ncbo/ontologies_api/issues/20
-            self.bring(:hasOntologyLanguage) unless self.loaded_attributes.include?(:hasOntologyLanguage)
-            cls_count = self.hasOntologyLanguage.skos? ? -1 : class_count(logger)
-            paging.page_count_set(cls_count) unless cls_count < 0
-            total_pages = paging.page(1, size).all.total_pages
-            num_threads = [total_pages, LinkedData.settings.indexing_num_threads].min
-            threads = []
-            page_classes = nil
-
-            num_threads.times do |num|
-              threads[num] = Thread.new {
-                Thread.current["done"] = false
-                Thread.current["page_len"] = -1
-                Thread.current["prev_page_len"] = -1
-
-                while !Thread.current["done"]
-                  synchronize do
-                    page = (page == 0 || page_classes.next?) ? page + 1 : nil
-
-                    if page.nil?
-                      Thread.current["done"] = true
-                    else
-                      Thread.current["page"] = page || "nil"
-                      RequestStore.store[:requested_lang] = :ALL
-                      page_classes = paging.page(page, size).all
-                      count_classes += page_classes.length
-                      Thread.current["page_classes"] = page_classes
-                      Thread.current["page_len"] = page_classes.length
-                      Thread.current["t0"] = Time.now
-
-                      # nothing retrieved even though we're expecting more records
-                      if total_pages > 0 && page_classes.empty? && (Thread.current["prev_page_len"] == -1 || Thread.current["prev_page_len"] == size)
-                        j = 0
-                        num_calls = LinkedData.settings.num_retries_4store
-
-                        while page_classes.empty? && j < num_calls do
-                          j += 1
-                          logger.error("Thread #{num + 1}: Empty page encountered. Retrying #{j} times...")
-                          sleep(2)
-                          page_classes = paging.page(page, size).all
-                          unless page_classes.empty?
-                            logger.info("Thread #{num + 1}: Success retrieving a page of #{page_classes.length} classes after retrying #{j} times...")
-                          end
-                        end
-
-                        if page_classes.empty?
-                          msg = "Thread #{num + 1}: Empty page #{Thread.current["page"]} of #{total_pages} persisted after retrying #{j} times. Indexing of #{self.id.to_s} aborted..."
-                          logger.error(msg)
-                          raise msg
-                        else
-                          Thread.current["page_classes"] = page_classes
-                        end
-                      end
-
-                      if page_classes.empty?
-                        if total_pages > 0
-                          logger.info("Thread #{num + 1}: The number of pages reported for #{self.id.to_s} - #{total_pages} is higher than expected #{page - 1}. Completing indexing...")
-                        else
-                          logger.info("Thread #{num + 1}: Ontology #{self.id.to_s} contains #{total_pages} pages...")
-                        end
-
-                        break
-                      end
-
-                      Thread.current["prev_page_len"] = Thread.current["page_len"]
-                    end
-                  end
-
-                  break if Thread.current["done"]
-
-                  logger.info("Thread #{num + 1}: Page #{Thread.current["page"]} of #{total_pages} - #{Thread.current["page_len"]} ontology terms retrieved in #{Time.now - Thread.current["t0"]} sec.")
-                  Thread.current["t0"] = Time.now
-
-                  Thread.current["page_classes"].each do |c|
-                    begin
-                      # this cal is needed for indexing of properties
-                      LinkedData::Models::Class.map_attributes(c, paging.equivalent_predicates, include_languages: true)
-                    rescue Exception => e
-                      i = 0
-                      num_calls = LinkedData.settings.num_retries_4store
-                      success = nil
-
-                      while success.nil? && i < num_calls do
-                        i += 1
-                        logger.error("Thread #{num + 1}: Exception while mapping attributes for #{c.id.to_s}. Retrying #{i} times...")
-                        sleep(2)
-
-                        begin
-                          LinkedData::Models::Class.map_attributes(c, paging.equivalent_predicates, include_languages: true)
-                          logger.info("Thread #{num + 1}: Success mapping attributes for #{c.id.to_s} after retrying #{i} times...")
-                          success = true
-                        rescue Exception => e1
-                          success = nil
-
-                          if i == num_calls
-                            logger.error("Thread #{num + 1}: Error mapping attributes for #{c.id.to_s}:")
-                            logger.error("Thread #{num + 1}: #{e1.class}: #{e1.message} after retrying #{i} times...\n#{e1.backtrace.join("\n\t")}")
-                            logger.flush
-                          end
-                        end
-                      end
-                    end
-
-                    synchronize do
-                      csv_writer.write_class(c)
-                    end
-                  end
-                  logger.info("Thread #{num + 1}: Page #{Thread.current["page"]} of #{total_pages} attributes mapped in #{Time.now - Thread.current["t0"]} sec.")
-
-                  Thread.current["t0"] = Time.now
-                  LinkedData::Models::Class.indexBatch(Thread.current["page_classes"])
-                  logger.info("Thread #{num + 1}: Page #{Thread.current["page"]} of #{total_pages} - #{Thread.current["page_len"]} ontology terms indexed in #{Time.now - Thread.current["t0"]} sec.")
-                  logger.flush
-                end
-              }
-            end
-
-            threads.map { |t| t.join }
-            csv_writer.close
-
-            begin
-              # index provisional classes
-              self.ontology.provisionalClasses.each { |pc| pc.index }
-            rescue Exception => e
-              logger.error("Error while indexing provisional classes for ontology #{self.ontology.acronym}:")
-              logger.error("#{e.class}: #{e.message}\n#{e.backtrace.join("\n\t")}")
-              logger.flush
-            end
-
-            if commit
-              t0 = Time.now
-              LinkedData::Models::Class.indexCommit()
-              logger.info("Ontology terms index commit in #{Time.now - t0} sec.")
-            end
-          rescue StandardError => e
-            csv_writer.close
-            logger.error("\n\n#{e.class}: #{e.message}\n")
-            logger.error(e.backtrace)
-            raise e
-          end
-        end
-        logger.info("Completed indexing ontology terms: #{self.ontology.acronym} in #{time} sec. #{count_classes} classes.")
-        logger.flush
-
-        if optimize
-          logger.info("Optimizing ontology terms index...")
-          time = Benchmark.realtime do
-            LinkedData::Models::Class.indexOptimize()
-          end
-          logger.info("Completed optimizing ontology terms index in #{time} sec.")
-        end
-      end
-
-      def index_properties(logger, commit = true, optimize = true)
-        page = 1
-        size = 2500
-        count_props = 0
-
-        time = Benchmark.realtime do
-          self.bring(:ontology) if self.bring?(:ontology)
-          self.ontology.bring(:acronym) if self.ontology.bring?(:acronym)
-          logger.info("Indexing ontology properties: #{self.ontology.acronym}...")
-          t0 = Time.now
-          self.ontology.unindex_properties(commit)
-          logger.info("Removed ontology properties index in #{Time.now - t0} seconds."); logger.flush
-
-          props = self.ontology.properties
-          count_props = props.length
-          total_pages = (count_props / size.to_f).ceil
-          logger.info("Indexing a total of #{total_pages} pages of #{size} properties each.")
-
-          props.each_slice(size) do |prop_batch|
-            t = Time.now
-            LinkedData::Models::OntologyProperty.indexBatch(prop_batch)
-            logger.info("Page #{page} of ontology properties indexed in #{Time.now - t} seconds."); logger.flush
-            page += 1
-          end
-
-          if commit
-            t0 = Time.now
-            LinkedData::Models::OntologyProperty.indexCommit(nil)
-            logger.info("Ontology properties index commit in #{Time.now - t0} seconds.")
-          end
-        end
-        logger.info("Completed indexing ontology properties of #{self.ontology.acronym} in #{time} sec. Total of #{count_props} properties indexed.")
-        logger.flush
-
-        if optimize
-          logger.info("Optimizing ontology properties index...")
-          time = Benchmark.realtime do
-            LinkedData::Models::OntologyProperty.indexOptimize(nil)
-          end
-          logger.info("Completed optimizing ontology properties index in #{time} seconds.")
-        end
-      end
-
       # Override delete to add removal from the search index
       # TODO: revise this with a better process
       def delete(*args)
@@ -1544,7 +632,7 @@ eos
           self.ontology.bring(:submissions)
 
           if self.ontology.submissions.length > 0
-            prev_sub = self.ontology.latest_submission()
+            prev_sub = self.ontology.latest_submission
 
             if prev_sub
               prev_sub.index_terms(LinkedData::Parser.logger || Logger.new($stderr))
@@ -1719,17 +807,6 @@ eos
         parsable
       end
 
-      private
-
-      def owlapi_parser_input
-        path = if zipped?
-                 self.zip_folder
-               else
-                 self.uploadFilePath
-               end
-        File.expand_path(path)
-      end
-
       def owlapi_parser(logger: Logger.new($stdout))
         unzip_submission(logger)
         LinkedData::Parser::OWLAPICommand.new(
@@ -1739,11 +816,15 @@ eos
           logger: logger)
       end
 
-      def delete_and_append(triples_file_path, logger, mime_type = nil)
-        Goo.sparql_data_client.delete_graph(self.id)
-        Goo.sparql_data_client.put_triples(self.id, triples_file_path, mime_type)
-        logger.info("Triples #{triples_file_path} appended in #{self.id.to_ntriples}")
-        logger.flush
+      private
+
+      def owlapi_parser_input
+        path = if zipped?
+                 self.zip_folder
+               else
+                 self.uploadFilePath
+               end
+        File.expand_path(path)
       end
 
       def check_http_file(url)
