+++ lib/ontologies_linked_data/services/submission_process/submission_process.rb
@@ -1,0 +1,18 @@
+module LinkedData
+  module Services
+    class OntologySubmissionProcess
+
+      def initialize(submission)
+        @submission = submission
+      end
+
+      def process(logger, options = {})
+        call
+      end
+
+      def call
+        raise NotImplementedError
+      end
+    end
+  end
+end
